src/backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-marshmallow
flask-cors
python-dotenv